templates/spec: refine sbindir evaluation
It originally erroneously checked for /sbin/weak-modules which could be
present or not in RHEL 7 and usually it didn't during the package build,
but this check succeeds once kmod package is installed, which breaks it.

Replacing it with more straightforward check for whether /sbin is a
directory and not a symbolic link. Hopefully it wouldn't break more than
it fixes.

* templates/spec (%define sbindir): Change test expression from
'-x "/sbin/weak-modules"' to '-d "/sbin" -a \! -h "/sbin"'.