ddiskit: change working directory to config's one
As most places expect that relative paths are relative to config
placement, this is a sane thing to do (at least, this is better than not
changing it).

* bin/ddiskit (parse_config): Call apply_config_file with filename
base name, as we are already in the directory where it is located.
(main): Change current working directory to the directory
where configuration file is located.