ddiskit: "Unicode fixes"
This is a bunch of changes that aims at using Unicode internally and
assuming that everything we get at input or put to output is UTF-8
encoded (even command line arguments, but come on, it's 2017 out there;
okay, not really, we try to guess encoding using sys.stding.encoding,
but fall back to "utf-8" in case it is not available).

* bin/ddiskit (import codecs): New imported module, for dealing with
decoding of input and encoding of output.
(DDiskitConfig.apply_args): New argument "encoding", defaults to
sys.stdin.encoding. Arguments are decoded with the encoding provided.
(cmd_prepare_sources, cnd_generate_spec, cmd_dump_config): Use
codecs.open() for dealing with file encoding.