templates/spec: split weak-modules call and initramfs generation
As noted in [1], latest change in a way weak-modules is called (that
deferrs the call to the %posttrans stage) is incompatible with the
latest changes in weak-modules itself. So, in order to fix this, we
revert back to calling weak-modules (albeit without initramfs
regeneration) in %post, and explicitly call dracut in %posttrans (in
order to satisfy original request [2]).

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1434128#c17
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1434128

* templates/spec (%post): Call weak-modules with --no-initramfs, touch
per-kernel-version status file for the initramfs regeneration in
%posttrans.
(%pretrans): Remove it.
(%posttrans): (Poorly) reimplement part of weak-modules and call dracut
for initramfs regeneration once for each kernel version.