templates/spec: add RHEL 8-specific kernel dependencies
RHEL 8 (somewhat ignorantly) changed the list of required build-time
dependencies for a kmod RPM.  This works this fact around in the ugliest
way possible, partly due to the fact that template has to work for all
supported OSes.

* ddiskit.config (spec_file): Add dist_build_dependencies configuration
option.
* templates/spec (kmod_dist_build_deps): New macro definition, derived
from spec_file.dist_build_dependencies configuration option.
[%if "%{kmod_dist_build_deps}" == ""] (kmod_dist_build_deps): Fall back
to the default of build-time dependencies based on RHEL/CentOS version.
[%if "%{kmod_dist_build_deps}" != ""]: Provide BuildRequires directive
with the value of the %{kmod_dist_build_deps} macro.

Signed-off-by: Eugene Syromiatnikov <esyr@redhat.com>